+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 2.0.0
 
 - Hidden units will be removed from new selections.
+- Refined orbital structure ambiguitiy with spawn layers.